requirements.txt
absl-py==2.0.0
apturl==0.5.2
astunparse==1.6.3
ayatana-settings==21.1.28
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
cajarename==21.11.24
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
deja-dup-caja==0.0.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
flatbuffers==23.5.26
folder-color-caja==0.0.86
folder-color-common==0.0.86
future==0.18.2
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.16.0
grpcio==1.59.2
h5py==3.10.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
joblib==1.3.2
keras==2.14.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
Magnus==1.0.3
Mako==1.1.3
Markdown==3.5.1
MarkupSafe==2.1.3
mate-hud==22.4.4
mate-tweak==22.4.8
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.3.0
protobuf==4.25.0
psutil==5.9.0
ptyprocess==0.7.0
pulsemixer==1.5.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pytest==7.4.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-xlib==0.29
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
ssh-import-id==5.11
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.5.3.1
tensorrt-dispatch==8.6.1
tensorrt-lean==8.6.1
termcolor==2.3.0
terminator==2.1.1
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
wrapt==1.14.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
